autoscheduler/requirements.txt
aiohttp==3.6.2
beautifulsoup4==4.8.2
discord.py==1.7.3
Django~=2.2.6
djangorestframework~=3.9.2
google-cloud-logging==2.3.1
lxml==4.5.0
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pylint==2.7.0
pylint-django==2.4.2
pytest==3.6
pytest-django==3.6.0
PyPDF2==1.26.0
requests==2.21.0
requests-mock==1.7.0
social-auth-app-django==3.1.0
python-dotenv==0.13.0
